move product picker component to own module
