+++ src/components/AdminPanel.js
@@ -1,32 +1,8 @@
 import React from 'react';
 import { Page, Card, Button, Layout } from '@shopify/polaris';
-import { ResourcePicker } from '@shopify/polaris/embedded';
 
 import SelectedProductListContainer from '../containers/SelectedProductListContainer'
-
-const ProductPicker = ({ open, onSelect, onToggle }) => {
-  return (
-    <ResourcePicker
-      products
-      allowMultiple
-      open={open}
-      onSelection={(resources) => {
-        const selectedProducts = resources.products.map((p) => {
-          return {
-            id: p.id,
-            title: p.title,
-            options: p.options,
-            variantCount: p.variants.length,
-            tags: p.tags,
-            variants: p.variants,
-          };
-        });
-        onSelect(selectedProducts);
-      }}
-      onCancel={() => onToggle()}
-    />
-  );
-}
+import ProductPicker from './ProductPicker';
 
 export default class AdminPanel extends React.Component {
   render() {
